requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
azure-core==1.30.2
azure-identity==1.17.1
black==24.8.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.29
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langsmith==0.1.98
marshmallow==3.21.3
msal==1.30.0
msal-extensions==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.40.3
orjson==3.10.7
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
portalocker==2.10.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
